+++ src/content_rules/actions/copy_item.ts
@@ -35,5 +35,10 @@ export const copy_item = {
     required: ['target_folder'],
     type: 'object',
   },
-  handler: async (params: Params, req: Request) => {},
+  handler: async (
+    params: Params,
+    document: any,
+    user: any,
+    contentRule: any,
+  ) => {},
 };