+++ src/content_rules/actions/delete_item.ts
@@ -21,5 +21,10 @@ export const delete_item = {
     required: [],
     type: 'object',
   },
-  handler: async (params: Params, req: Request) => {},
+  handler: async (
+    params: Params,
+    document: any,
+    user: any,
+    contentRule: any,
+  ) => {},
 };