+++ src/content_rules/conditions/content_type.ts
@@ -48,5 +48,12 @@ export const content_type = {
     required: ['check_types'],
     type: 'object',
   },
-  handler: async (params: Params, req: Request) => {},
+  handler: async (
+    params: Params,
+    document: any,
+    user: any,
+    contentRule: any,
+  ) => {
+    return true;
+  },
 };