+++ src/events/cache/cache.ts
@@ -6,44 +6,59 @@
 import { purgeXkey } from '../../helpers/cache/cache';
 
 export const purge = {
-  onAfterModified: async (context: any) => {
-    purgeXkey(context.uuid);
+  onAfterModified: async (document: any) => {
+    purgeXkey(document.uuid);
   },
 
-  onBeforeDelete: async (context: any) => {
-    purgeXkey(context.uuid);
+  onBeforeDelete: async (document: any) => {
+    purgeXkey(document.uuid);
   },
 
-  onAfterMove: async (context: any) => {
-    purgeXkey(context.uuid);
+  onAfterMove: async (document: any) => {
+    purgeXkey(document.uuid);
   },
 
-  onAfterDeleteUser: async (userid: string) => {
-    purgeXkey(userid);
-    purgeXkey('users');
-  },
-
-  onAfterUpdateUser: async (user: any) => {
+  onAfterDeleteUser: async (_document: any, user: any) => {
     purgeXkey(user.id);
     purgeXkey('users');
   },
 
-  onAfterAddUser: async (user: any) => {
+  onAfterUpdateUser: async (_document: any, user: any) => {
     purgeXkey(user.id);
     purgeXkey('users');
   },
 
-  onAfterDeleteGroup: async (groupid: string) => {
-    purgeXkey(groupid);
-    purgeXkey('groups');
+  onAfterAddUser: async (_document: any, user: any) => {
+    purgeXkey(user.id);
+    purgeXkey('users');
   },
 
-  onAfterUpdateGroup: async (group: any) => {
+  onAfterDeleteGroup: async (
+    _document: any,
+    _user: any,
+    _trx: any,
+    group: any,
+  ) => {
     purgeXkey(group.id);
     purgeXkey('groups');
   },
 
-  onAfterAddGroup: async (group: any) => {
+  onAfterUpdateGroup: async (
+    _document: any,
+    _user: any,
+    _trx: any,
+    group: any,
+  ) => {
+    purgeXkey(group.id);
+    purgeXkey('groups');
+  },
+
+  onAfterAddGroup: async (
+    _document: any,
+    _user: any,
+    _trx: any,
+    group: any,
+  ) => {
     purgeXkey(group.id);
     purgeXkey('groups');
   },